<refactor>

+++ src/ru/yandex/practicum/tasktracker/manger/HistoryManager.java
@@ -8,5 +8,5 @@ public interface HistoryManager {
 
     void add(Task task);
 
-    void remove(long id);
+    void remove(int id);
 }
